Factor out a membership mixin
