+++ gratipay/models/team/mixins/__init__.py
@@ -1,5 +1,6 @@
 from .available import AvailableMixin as Available
 from .distributing import DistributingMixin as Distributing
+from .membership import MembershipMixin as Membership
 from .tip_migration import TipMigrationMixin as TipMigration
 
-__all__ = ['Available', 'Distributing', 'TipMigration']
+__all__ = ['Available', 'Distributing', 'Membership', 'TipMigration']
